requirements.txt
pydantic>=2.0.0
rich>=10.0.0
requests>=2.25.1
python-dotenv>=0.19.0
pandas>=1.3.0